+++ psRadiantOne/Public/NamingContexts/New-R1NamingContextLabel.ps1
@@ -1,0 +1,43 @@
+# .ExternalHelp psRadiantOne-help.xml
+function New-R1NamingContextLabel {
+	[CmdletBinding(SupportsShouldProcess, ConfirmImpact = 'Medium')]
+	[OutputType([void])]
+	param(
+		[parameter(
+			Mandatory = $true,
+			ValueFromPipelineByPropertyName = $true
+		)]
+		[ValidateNotNullOrEmpty()]
+		[string]$dn,
+
+		[parameter(
+			Mandatory = $true,
+			ValueFromPipelineByPropertyName = $true
+		)]
+		[ValidateNotNullOrEmpty()]
+		[string]$rdn
+	)
+
+	Begin {
+
+		Assert-R1Session -RequireToken
+
+	}#begin
+
+	Process {
+
+		$URI = Resolve-R1ServiceUrl -Service Namespace -Path "naming_contexts/$($dn | Get-EscapedString)/add_label"
+
+		$Body = @{ rdn = $rdn } | ConvertTo-R1JsonBody
+
+		if ($PSCmdlet.ShouldProcess($dn, "Add Child Label '$rdn'")) {
+
+			$null = Invoke-R1RestMethod -Uri $URI -Method POST -Body $Body
+
+		}
+
+	}#process
+
+	End { }#end
+
+}